langgraph-runner/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==1.59.3
pydantic==2.10.3
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
